+++ lib/src/main/java/com/smileidentity/networking/NetworkingUtil.kt
@@ -10,6 +10,7 @@ import java.io.BufferedOutputStream
 import java.io.File
 import java.io.FileOutputStream
 import java.util.zip.ZipEntry
+import java.util.zip.ZipException
 import java.util.zip.ZipOutputStream
 
 fun calculateSignature(timestamp: String): String {
@@ -24,16 +25,17 @@ fun calculateSignature(timestamp: String): String {
 }
 
 fun UploadRequest.zip(): File {
+    val uploadRequest = deDupedUploadRequest(this)
     val zipFile = File.createTempFile("upload", ".zip")
     val zipOutputStream = ZipOutputStream(BufferedOutputStream(FileOutputStream(zipFile)))
 
     // Write info.json
     zipOutputStream.putNextEntry(ZipEntry("info.json"))
-    zipOutputStream.write(moshi.adapter(UploadRequest::class.java).toJson(this).toByteArray())
+    zipOutputStream.write(moshi.adapter(UploadRequest::class.java).toJson(uploadRequest).toByteArray())
     zipOutputStream.closeEntry()
 
     // Write images
-    this.images.forEach { imageInfo ->
+    uploadRequest.images.forEach { imageInfo ->
         zipOutputStream.putNextEntry(ZipEntry(imageInfo.image.name))
         imageInfo.image.inputStream().use { it.copyTo(zipOutputStream) }
         zipOutputStream.closeEntry()
@@ -43,6 +45,19 @@ fun UploadRequest.zip(): File {
     zipFile.deleteOnExit()
     return zipFile
 }
+
+/**
+ * This function takes an [UploadRequest] and returns a new [UploadRequest] with duplicate images
+ * removed. This is necessary because we can't have duplicate entries in a zip file. There is no
+ * valid use-case where we would want to upload the same image twice. The only scenario where that
+ * could happen is if the user selects the same image for both the front and back of a document.
+ *
+ * If we don't do this, then we will crash when attempting to include the file in the Zip with a
+ * [ZipException] stating that the entry already exists
+ */
+private fun deDupedUploadRequest(uploadRequest: UploadRequest) = uploadRequest.copy(
+    images = uploadRequest.images.distinctBy { it.image.name },
+)
 
 fun File.asSelfieImage() = UploadImageInfo(
     imageTypeId = ImageType.SelfieJpgFile,